disk: boxed large sector data that caused segfault
Allocating a Disk on the stack caused a stack overflow which turned into
a segfault. This was because the sector field was overflowwing the
stack. Boxing the offending data now causes it to be allocated on the
heap which turns `sectors` into Option<pointer>, which fits on the stack
nicely.